feat: add links to tag list
