+++ vite-project/src/components/List/PostList.tsx
@@ -7,13 +7,20 @@ interface IProps {
   postList: IPost[],
   openForm: (post?: IPost) => void,
   deletePost: (id: number) => void
+  setSearch: React.Dispatch<React.SetStateAction<string>>
 }
 
-const PostList = ({postList, openForm, deletePost}:IProps) => {
+const PostList = ({postList, openForm, deletePost, setSearch}:IProps) => {
   const result = postList.length
   ? (<div className={classes.postlist}>
       {postList.map((el) => 
-        <Post key={el.id} post={el} deletePost={deletePost} openForm={openForm} />
+        <Post 
+          key={el.id} 
+          post={el} 
+          deletePost={deletePost} 
+          openForm={openForm}
+          setSearch={setSearch} 
+        />
       )}
     </div>)
   : <EmptyList />
