+++ vite-project/src/components/TagList/TagList.tsx
@@ -1,0 +1,29 @@
+import classes from "./TagList.module.scss"
+
+interface IProps {
+  tags: string[],
+  setSearch: React.Dispatch<React.SetStateAction<string>>,
+  [key: string]: unknown
+}
+
+const TagList = ({tags, setSearch, ...props}:IProps) => {
+  const result = tags.length
+  ? (
+    <div {...props}>
+      {tags.map((el) => 
+        <span 
+          key={el} 
+          className={classes.taglink}
+          onClick={() => setSearch(el)}
+        >
+          {el}
+        </span>
+      )}
+    </div>
+    )
+  : <div>No tags</div>;
+
+  return result;
+};
+
+export default TagList;
